correctly handle requirements with square brackets
When encountering a requirement like `psycopg[binary,pool]==3.2.9` or
`python-jose[cryptography]==3.4.0` it was treating the entire left
half as the package name to look up in PyPI, which would then fail and
it would just ignore that one, missing any potential updates. This
tightens up the name matching so it would find `psycopg` or
`python-jose` respectively and be able to look up the correct package
in PyPI. Ideally, it would maintain the full name string with the
optional dependencies and check those as well. For now, this is a
quick improvement that at least makes it work correctly for the base
package.